Fix: 컴포넌트 prop 으로 속성제어하도록 변경
